example_proj/requirements.txt
Django
psycopg2
djangorestframework
python-dotenv

